pyproject.toml
[project]
name = "flet-app-template"
version = "0.1.0"
description = "A template for building desktop applications with Flet."
authors = [{name = "Valmisson Grizorte <valmisson.github.io>"}]
readme = "README.md"
requires-python = ">=3.11,<4.0"

[tool.poetry.dependencies]
flet = "^0.28.3"
sentry-sdk = "^2.37.0"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
mypy = "^1.17.1"
pre-commit = "^4.3.0"
ruff = "^0.12.12"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
